document-processing/requirements.txt
Flask==3.0.0
python-docx==1.1.0
Werkzeug==3.0.1
openai>=2.0.0
httpx==0.27.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3
